# internal/modules/message/message.controller.go
package message

import (
	"chat-server/internal/middleware"
	"chat-server/internal/utils"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"go.uber.org/zap"
)

type Controller struct {
	service *Service
	logger  *zap.SugaredLogger
}

func NewController(service *Service, logger *zap.SugaredLogger) *Controller {
	return &Controller{
		service: service,
		logger:  logger.Named("[message_controller]"),
	}
}

func (ctrl *Controller) RegisterRoutes(router *utils.AppGroup, authMiddleware *middleware.AuthMiddleware) {
	messages := router.Group("/messages")
	messages.Use(authMiddleware.RequireAuth())
	{
		messages.POST("", ctrl.SendMessage)
		messages.GET("/:conversationId", ctrl.GetMessages)
		messages.DELETE("/:conversationId/:messageId", ctrl.DeleteMessage)
	}
}

// SendMessage godoc
// @Summary      Send message
// @Description  Send a message in a conversation
// @Tags         messages
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request body SendMessageRequest true "Send Message"
// @Success      201  {object}  MessageSuccessResponse
// @Failure      400  {object}  utils.APIError
// @Failure      401  {object}  utils.APIError
// @Failure      403  {object}  utils.APIError
// @Failure      404  {object}  utils.APIError
// @Router       /messages [post]
func (ctrl *Controller) SendMessage(c *gin.Context) (interface{}, error) {
	userID, exists := middleware.GetUserID(c)
	if !exists {
		return nil, utils.NewHTTPError(http.StatusUnauthorized, "user not authenticated")
	}

	var req SendMessageRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		return nil, utils.NewHTTPError(http.StatusBadRequest, "invalid request body")
	}

	conversationID, err := uuid.Parse(req.ConversationID)
	if err != nil {
		return nil, utils.NewHTTPError(http.StatusBadRequest, "invalid conversation ID")
	}

	var replyToID *uuid.UUID
	if req.ReplyToID != nil && *req.ReplyToID != "" {
		parsed, err := uuid.Parse(*req.ReplyToID)
		if err != nil {
			return nil, utils.NewHTTPError(http.StatusBadRequest, "invalid reply to ID")
		}
		replyToID = &parsed
	}

	message, err := ctrl.service.SendMessage(userID, conversationID, req.Type, req.Content, req.Metadata, replyToID)
	if err != nil {
		ctrl.logger.Errorw("Failed to send message", "error", err)
		return nil, utils.NewHTTPError(http.StatusInternalServerError, "failed to send message")
	}

	return message, nil
}

// GetMessages godoc
// @Summary      Get messages
// @Description  Get messages from a conversation with pagination
// @Tags         messages
// @Produce      json
// @Security     BearerAuth
// @Param        conversationId path string true "Conversation ID"
// @Param        limit query int false "Limit number of messages" default(50)
// @Param        before query string false "Get messages before this message ID (timeuuid)"
// @Success      200  {object}  MessagesListSuccessResponse
// @Failure      400  {object}  utils.APIError
// @Failure      401  {object}  utils.APIError
// @Failure      403  {object}  utils.APIError
// @Failure      404  {object}  utils.APIError
// @Router       /messages/{conversationId} [get]
func (ctrl *Controller) GetMessages(c *gin.Context) (interface{}, error) {
	userID, exists := middleware.GetUserID(c)
	if !exists {
		return nil, utils.NewHTTPError(http.StatusUnauthorized, "user not authenticated")
	}

	conversationID, err := uuid.Parse(c.Param("conversationId"))
	if err != nil {
		return nil, utils.NewHTTPError(http.StatusBadRequest, "invalid conversation ID")
	}

	limit := 50
	if limitStr := c.Query("limit"); limitStr != "" {
		if parsedLimit, err := strconv.Atoi(limitStr); err == nil && parsedLimit > 0 {
			limit = parsedLimit
		}
	}

	var beforeMessageID *string
	if before := c.Query("before"); before != "" {
		beforeMessageID = &before
	}

	messages, err := ctrl.service.GetMessages(userID, conversationID, limit, beforeMessageID)
	if err != nil {
		ctrl.logger.Errorw("Failed to get messages", "error", err)
		return nil, utils.NewHTTPError(http.StatusInternalServerError, "failed to get messages")
	}

	return messages, nil
}

// DeleteMessage godoc
// @Summary      Delete message
// @Description  Delete a message (soft delete)
// @Tags         messages
// @Produce      json
// @Security     BearerAuth
// @Param        conversationId path string true "Conversation ID"
// @Param        messageId path string true "Message ID (timeuuid)"
// @Success      200  {object}  SimpleSuccessResponse
// @Failure      400  {object}  utils.APIError
// @Failure      401  {object}  utils.APIError
// @Failure      403  {object}  utils.APIError
// @Failure      404  {object}  utils.APIError
// @Router       /messages/{conversationId}/{messageId} [delete]
func (ctrl *Controller) DeleteMessage(c *gin.Context) (interface{}, error) {
	userID, exists := middleware.GetUserID(c)
	if !exists {
		return nil, utils.NewHTTPError(http.StatusUnauthorized, "user not authenticated")
	}

	conversationID := c.Param("conversationId")
	messageID := c.Param("messageId")

	if err := ctrl.service.DeleteMessage(userID, conversationID, messageID); err != nil {
		ctrl.logger.Errorw("Failed to delete message", "error", err)
		return nil, utils.NewHTTPError(http.StatusInternalServerError, "failed to delete message")
	}

	return nil, nil
}